+++ noodles-csi/src/binning_index/index/indexer.rs
@@ -1,5 +1,6 @@
-use std::{io, mem};
+use std::io;
 
+use indexmap::IndexMap;
 use noodles_core::Position;
 
 use super::{
@@ -13,7 +14,6 @@ pub struct Indexer<I> {
     min_shift: u8,
     depth: u8,
     header: Option<Header>,
-    reference_sequence_builder: reference_sequence::Builder<I>,
     reference_sequences: Vec<ReferenceSequence<I>>,
     unplaced_unmapped_record_count: u64,
 }
@@ -35,7 +35,6 @@ where
             min_shift,
             depth,
             header: None,
-            reference_sequence_builder: reference_sequence::Builder::default(),
             reference_sequences: Vec::new(),
             unplaced_unmapped_record_count: 0,
         }
@@ -99,7 +98,14 @@ where
             return Ok(());
         };
 
-        match reference_sequence_id.cmp(&self.current_reference_sequence_id()) {
+        if self.reference_sequences.is_empty() {
+            self.add_reference_sequences_until(0);
+        }
+
+        // SAFETY: `self.reference_sequences` is non-empty.
+        let current_reference_sequence_id = self.reference_sequences.len() - 1;
+
+        match reference_sequence_id.cmp(&current_reference_sequence_id) {
             Ordering::Less => {
                 return Err(io::Error::new(
                     io::ErrorKind::InvalidInput,
@@ -107,17 +113,11 @@ where
                 ));
             }
             Ordering::Equal => {}
-            Ordering::Greater => self.add_reference_sequences_builders_until(reference_sequence_id),
+            Ordering::Greater => self.add_reference_sequences_until(reference_sequence_id),
         }
 
-        self.reference_sequence_builder.add_record(
-            self.min_shift,
-            self.depth,
-            start,
-            end,
-            is_mapped,
-            chunk,
-        );
+        let reference_sequence = &mut self.reference_sequences[reference_sequence_id];
+        reference_sequence.update(self.min_shift, self.depth, start, end, is_mapped, chunk);
 
         Ok(())
     }
@@ -138,7 +138,8 @@ where
                 .build();
         }
 
-        self.add_reference_sequences_builders_until(reference_sequence_count);
+        // SAFETY: `reference_sequence_count` is > 0.
+        self.add_reference_sequences_until(reference_sequence_count - 1);
 
         let mut builder = Index::builder()
             .set_reference_sequences(self.reference_sequences)
@@ -151,16 +152,12 @@ where
         builder.build()
     }
 
-    fn current_reference_sequence_id(&self) -> usize {
-        self.reference_sequences.len()
-    }
+    fn add_reference_sequences_until(&mut self, reference_sequence_id: usize) {
+        let new_len = reference_sequence_id + 1;
 
-    fn add_reference_sequences_builders_until(&mut self, reference_sequence_id: usize) {
-        while self.reference_sequences.len() < reference_sequence_id {
-            let reference_sequence_builder = mem::take(&mut self.reference_sequence_builder);
-            let reference_sequence = reference_sequence_builder.build();
-            self.reference_sequences.push(reference_sequence);
-        }
+        self.reference_sequences.resize_with(new_len, || {
+            ReferenceSequence::new(IndexMap::new(), Default::default(), None)
+        });
     }
 }
 
@@ -173,7 +170,6 @@ where
             min_shift: 14,
             depth: 5,
             header: None,
-            reference_sequence_builder: reference_sequence::Builder::default(),
             reference_sequences: Vec::new(),
             unplaced_unmapped_record_count: 0,
         }
@@ -182,12 +178,14 @@ where
 
 #[cfg(test)]
 mod tests {
+    use noodles_bgzf as bgzf;
+
     use super::*;
-    use crate::binning_index::index::reference_sequence::index::BinnedIndex;
+    use crate::binning_index::index::reference_sequence::{index::LinearIndex, Bin, Metadata};
 
     #[test]
     fn test_default() {
-        let indexer = Indexer::<BinnedIndex>::default();
+        let indexer = Indexer::<LinearIndex>::default();
 
         assert_eq!(indexer.min_shift, 14);
         assert_eq!(indexer.depth, 5);
@@ -197,8 +195,91 @@ mod tests {
     }
 
     #[test]
-    fn test_build() {
-        let index = Indexer::<BinnedIndex>::default().build(2);
+    fn test_build() -> Result<(), Box<dyn std::error::Error>> {
+        const MIN_SHIFT: u8 = 14;
+        const DEPTH: u8 = 5;
+
+        let mut indexer = Indexer::<LinearIndex>::new(MIN_SHIFT, DEPTH);
+
+        indexer.add_record(
+            Some((0, Position::try_from(8)?, Position::try_from(13)?, true)),
+            Chunk::new(
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(9),
+            ),
+        )?;
+
+        indexer.add_record(
+            Some((
+                0,
+                Position::try_from(121393)?,
+                Position::try_from(196418)?,
+                false,
+            )),
+            Chunk::new(
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(3473408),
+            ),
+        )?;
+
+        let actual = indexer.build(1);
+
+        let expected = {
+            let bins = [
+                (
+                    4681,
+                    Bin::new(vec![Chunk::new(
+                        bgzf::VirtualPosition::from(0),
+                        bgzf::VirtualPosition::from(9),
+                    )]),
+                ),
+                (
+                    73,
+                    Bin::new(vec![Chunk::new(
+                        bgzf::VirtualPosition::from(9),
+                        bgzf::VirtualPosition::from(3473408),
+                    )]),
+                ),
+            ]
+            .into_iter()
+            .collect();
+
+            let index = vec![
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+            ];
+
+            let metadata = Metadata::new(
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(3473408),
+                1,
+                1,
+            );
+
+            Index::builder()
+                .set_reference_sequences(vec![ReferenceSequence::new(bins, index, Some(metadata))])
+                .set_unplaced_unmapped_record_count(0)
+                .build()
+        };
+
+        assert_eq!(actual, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_build_with_reference_sequence_count() {
+        let index = Indexer::<LinearIndex>::default().build(2);
         assert_eq!(index.reference_sequences().len(), 2);
     }
 }